Update test to reflect moved code
